+++ package.json
@@ -23,22 +23,10 @@
     "karma-phantomjs-launcher": "1.0.0",
     "karma-safari-launcher": "0.1.1",
     "lodash": "4.6.1",
-    "main-bower-files": "2.11.1",
     "merge2": "1.0.1",
     "phantomjs-prebuilt": "2.1.7",
     "run-sequence": "1.1.5",
     "socket.io": "1.3.7",
     "through2": "2.0.1",
-    "web-tools": "git+https://github.com/imgix/web-tools.git#7.1.0"
-  },
-  "devDependencies": {
-    "gulp": "3.9.1",
-    "gulp-jscs": "3.0.2",
-    "gulp-jshint": "2.0.0",
-    "gulp-util": "3.0.7",
-    "jshint": "2.9.1",
-    "lodash": "4.6.1",
-    "reporter-plus": "git://github.com/jayeb/reporter-plus.git",
-    "yargs": "4.2.0"
   }
 }
